Множество изменений в положении классов
