+++ start_screen.py
@@ -3,6 +3,8 @@ import sys
 
 import pygame
 
+from our_tools import load_image
+
 FPS = 50
 pygame.init()
 size = WIDTH, HEIGHT = 800, 500
@@ -10,23 +12,6 @@ screen = pygame.display.set_mode(size)
 pygame.display.set_caption('PyCharm Survivors')
 
 clock = pygame.time.Clock()
-
-
-def load_image(name, colorkey=None):
-    fullname = os.path.join(name)
-    # если файл не существует, то выходим
-    if not os.path.isfile(fullname):
-        print(f"Файл с изображением '{fullname}' не найден")
-        sys.exit()
-    image = pygame.image.load(fullname)
-    if colorkey is not None:
-        image = image.convert()
-        if colorkey == -1:
-            colorkey = image.get_at((0, 0))
-        image.set_colorkey(colorkey)
-    else:
-        image = image.convert_alpha()
-    return image
 
 
 def terminate():
@@ -37,7 +22,7 @@ def terminate():
 def start_screen():
     intro_text = ['']
     #
-    fon = pygame.transform.scale(load_image('data/logo-2.jpg'), (WIDTH, HEIGHT))
+    fon = pygame.transform.scale(load_image('logo-2.jpg'), (WIDTH, HEIGHT))
     screen.blit(fon, (0, 0))
     font = pygame.font.Font(None, 30)
     text_coord = 150
